Recomputed elastic scattering tables with two changes
Change #1 is to the interpolation over energy. Previous tables started
with a coarse but logarithmically uniform sampling of energies after
which intervals were subdivided as needed to reach tolerance. The new
tables start from a manually provided set of energies that includes
round-numbered values that are frequently used. These intervals are
likewise subdivided as needed.

Change #2 is to the interpolation over r (random number in [0,1]). The
new tables have a tighter error tolerance that has resulted in a larger
number of r-samples.