Symmetrical treatment of k1 and k2
The change is to the high-k part of the algorithm for finite barrier
widths. The full formula has numerical imprecision if k (electron wave
vector) is too high. The previous algorithm only tested k1, the initial
wave vector. We now check both. We also reduce the threshold at which we
switch to an approximate formula that does not suffer from the
imprecision but which is valid only for k1 or k2 large enough.