Added methods to permit valence band reference for barrier scattering
ExpQMBarrierSM now allows specification of an energy offset. Electron
kinetic energy is treated as larger by this amount. If the offset is
specified as the energy difference between conduction and valence band
bottoms, it has the effect of referencing electron energies to the
bottom of the valence band for the purpose of barrier scattering.

SEmaterial now allows direct specification of the surface potential
energy barrier. If it is unspecified the barrier defaults to the
difference between vacuum energy and conduction band bottom. However, to
use a valence band reference one can specify the barrier to be larger by
the same energy offset used in ExpQMBarrierSM.